LayoutTests/fast/files/file-reader.html

<!DOCTYPE html>
<html>
<body>
<input type="file" name="file" id="file" onchange="onInputFileChange()" multiple>
<pre id='console'></pre>

<script>
var testCases = [
    "testReadingNonExistentFile",
    "testReadingEmptyFileAsBinaryString",
    "testReadingEmptyFileAsText",
    "testReadingEmptyFileAsDataURL",
    "testReadingUTF8EncodedFileAsBinaryString",
    "testReadingUTF8EncodedFileAsText",
    "testReadingUTF16BEBOMEncodedFileAsText",
    "testReadingUTF16LEBOMEncodedFileAsText",
    "testReadingUTF8BOMEncodedFileAsText",
    "testReadingUTF16BEEncodedFileAsTextWithUTF16Encoding",
    "testReadingUTF8EncodedFileAsDataURL",
    "testMultipleReads",
];
var testIndex = 0;

var testFileInfoList = [
    { 'name': 'non-existent', 'path': 'resources/non-existent' },
    { 'name': 'empty-file', 'path': 'resources/empty-file' },
    { 'name': 'UTF8-file', 'path': 'resources/UTF8.txt' },
    { 'name': 'UTF16BE-BOM-file', 'path': 'resources/UTF16BE-BOM.txt' },
    { 'name': 'UTF16LE-BOM-file', 'path': 'resources/UTF16LE-BOM.txt' },
    { 'name': 'UTF8-BOM-file', 'path': 'resources/UTF8-BOM.txt' },
    { 'name': 'UTF16BE-file', 'path': 'resources/UTF16BE.txt' },
];
var testFiles = { };

function log(message)
{
    document.getElementById('console').appendChild(document.createTextNode(message + "\n"));
}

function createFileReader()
{
    var reader = new FileReader();

    reader.onloadstart = loadStarted;
    reader.onload = loaded;
    reader.onabort = logEvent;
    reader.onloadend = loadEnded;
    reader.onerror = loadFailed;

    log("readyState: " + reader.readyState);
    return reader;
}

function logEvent(event)
{
    log("Received " + event.type + " event");
}

function loadStarted(event)
{
    logEvent(event);
    log("readyState: " + event.target.readyState);
}

function loaded(event)
{
    logEvent(event);
    log("readyState: " + event.target.readyState);
    log("result size: " + event.target.result.length);
    log("result: " + event.target.result);
}

function loadFailed(event)
{
    logEvent(event);
    log("readyState: " + event.target.readyState);
    log("error code: " + event.target.error.code);
}

function loadEnded(event)
{
    logEvent(event);
    runNextTest();
}

function onInputFileChange()
{
    var files = document.getElementById("file").files;
    for (var i = 0; i < files.length; i++)
        testFiles[testFileInfoList[i]['name']] = files[i];

    runNextTest();
}

function runNextTest()
{
    if (testIndex < testCases.length) {
        testIndex++;
        window[testCases[testIndex - 1]]();
    } else {
        log("DONE");
        if (window.layoutTestController)
            layoutTestController.notifyDone();
    }
}

function testReadingNonExistentFile()
{
    log("Test reading a non-existent file");
    var reader = createFileReader();
    reader.readAsBinaryString(testFiles['non-existent']);
}

function testReadingEmptyFileAsBinaryString()
{
    log("Test reading an empty file as binary string");
    var reader = createFileReader();
    reader.readAsBinaryString(testFiles['empty-file']);
}

function testReadingEmptyFileAsText()
{
    log("Test reading an empty file as text");
    var reader = createFileReader();
    reader.readAsText(testFiles['empty-file']);
}

function testReadingEmptyFileAsDataURL()
{
    log("Test reading an empty file as data URL");
    var reader = createFileReader();
    reader.readAsDataURL(testFiles['empty-file']);
}

function testReadingUTF8EncodedFileAsBinaryString()
{
    log("Test reading a UTF-8 file as binary string");
    var reader = createFileReader();
    reader.readAsBinaryString(testFiles['UTF8-file']);
}

function testReadingUTF8EncodedFileAsText()
{
    log("Test reading a UTF-8 file as text");
    var reader = createFileReader();
    reader.readAsText(testFiles['UTF8-file']);
}

function testReadingUTF16BEBOMEncodedFileAsText()
{
    log("Test reading a UTF-16BE BOM file as text");
    var reader = createFileReader();
    reader.readAsText(testFiles['UTF16BE-BOM-file']);
}

function testReadingUTF16LEBOMEncodedFileAsText()
{
    log("Test reading a UTF-16LE BOM file as text");
    var reader = createFileReader();
    reader.readAsText(testFiles['UTF16LE-BOM-file']);
}

function testReadingUTF8BOMEncodedFileAsText()
{
    log("Test reading a UTF-8 BOM file as text");
    var reader = createFileReader();
    reader.readAsText(testFiles['UTF8-BOM-file']);
}

function testReadingUTF16BEEncodedFileAsTextWithUTF16Encoding()
{
    log("Test reading a UTF-16BE file as text with UTF-16BE encoding");
    var reader = createFileReader();
    reader.readAsText(testFiles['UTF16BE-file'], "UTF-16BE");
}

function testReadingUTF8EncodedFileAsDataURL()
{
    log("Test reading a UTF-8 file as data URL");
    var reader = createFileReader();
    reader.readAsDataURL(testFiles['UTF8-file']);
}

function testMultipleReads()
{
    log("Test calling multiple read methods and only last one is processed");
    var reader = createFileReader();
    reader.readAsBinaryString(testFiles['UTF8-file']);
    reader.readAsText(testFiles['UTF8-file']);
    reader.readAsDataURL(testFiles['UTF8-file']);
}

function runTests()
{
    var pathsOnly = testFileInfoList.map(function(fileSpec) { return fileSpec['path']; });
    eventSender.beginDragWithFiles(pathsOnly);
    eventSender.mouseMoveTo(10, 10);
    eventSender.mouseUp();
}

if (window.eventSender) {
    layoutTestController.dumpAsText();
    layoutTestController.waitUntilDone();
    window.onload = runTests;
}
</script>
</body>
</html>
