LayoutTests/fast/files/file-reader-expected.txt


Test reading a non-existent file
readyState: 0
Received error event
readyState: 2
error code: 8
Received loadend event
Test reading an empty file as binary string
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 0
result: 
Received loadend event
Test reading an empty file as text
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 0
result: 
Received loadend event
Test reading an empty file as data URL
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: data:
Received loadend event
Test reading a UTF-8 file as binary string
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-8 file as text
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-16BE BOM file as text
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-16LE BOM file as text
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-8 BOM file as text
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-16BE file as text with UTF-16BE encoding
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 5
result: Hello
Received loadend event
Test reading a UTF-8 file as data URL
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 31
result: data:text/plain;base64,SGVsbG8=
Received loadend event
Test calling multiple read methods and only last one is processed
readyState: 0
Received loadstart event
readyState: 1
Received load event
readyState: 2
result size: 31
result: data:text/plain;base64,SGVsbG8=
Received loadend event
DONE
